fix(web3): `GwConfig` is not a constructor
For `EthRegistryAddress` and `GwConfig` circular references
