+++ web3/packages/api-server/src/base/address.ts
@@ -1,54 +1,21 @@
 import { Hash, HexString, Script, utils } from "@ckb-lumos/base";
 import { GodwokenClient } from "@godwoken-web3/godwoken";
 import { Store } from "../cache/store";
+import { BaseEthRegistryAddress } from "./base-eth-registry-address";
 import { gwConfig } from "./index";
 import { logger } from "./logger";
-import { Uint32 } from "./types/uint";
 
 // the eth address vs script hash is not changeable, so we set no expire for cache
 const scriptHashCache = new Store(false);
 
 // Only support eth address now!
-export class EthRegistryAddress {
-  private registryId: number;
-  private addressByteSize: number = 20;
-  public readonly address: HexString;
-
+export class EthRegistryAddress extends BaseEthRegistryAddress {
   // Using optional registryId when gwConfig not initialized
   constructor(
     address: HexString,
     { registryId }: { registryId?: number } = {}
   ) {
-    if (!address.startsWith("0x") || address.length !== 42) {
-      throw new Error(`Eth address format error: ${address}`);
-    }
-    this.address = address.toLowerCase();
-
-    this.registryId = registryId || +gwConfig.accounts.ethAddrReg.id;
-  }
-
-  public serialize(): HexString {
-    return (
-      "0x" +
-      new Uint32(this.registryId).toLittleEndian().slice(2) +
-      new Uint32(this.addressByteSize).toLittleEndian().slice(2) +
-      this.address.slice(2)
-    );
-  }
-
-  public static Deserialize(hex: HexString): EthRegistryAddress {
-    const hexWithoutPrefix = hex.slice(2);
-    const registryId: number = Uint32.fromLittleEndian(
-      hexWithoutPrefix.slice(0, 8)
-    ).getValue();
-    const addressByteSize: number = Uint32.fromLittleEndian(
-      hexWithoutPrefix.slice(8, 16)
-    ).getValue();
-    const address: HexString = hexWithoutPrefix.slice(16);
-    if (addressByteSize !== 20 || address.length !== 40) {
-      throw new Error(`Eth address deserialize error: ${hex}`);
-    }
-    return new EthRegistryAddress("0x" + address, { registryId });
+    super(address, registryId || +gwConfig.accounts.ethAddrReg.id);
   }
 }
